리팩토링: 재사용을 위해 각 value 별로 Form component 분리
