+++ src/components/form/StarRating.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StarRating = () => {
+  return <Wrap></Wrap>;
+};
+
+export default StarRating;
+
+const Wrap = styled.div``;
