+++ src/pages/reviewForm/ReviewForm.js
@@ -1,10 +1,15 @@
-import React, { useState, useRef } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import VegLevel from './VegLevel';
+import Category from '../../components/form/Category';
+import UploadImg from '../../components/form/UploadImg';
+import VegLevel from '../../components/form/VegLevel';
+import Input from '../../components/form/Input';
+import Textarea from '../../components/form/Textarea';
+import SubmitBtn from '../../components/form/SubmitBtn';
 import BgImg from './images/pexels-karolina-grabowska-4197908.jpg';
 
 const initReview = {
-  category: '',
+  category: 'store',
   previewImgs: [],
   vegLevel: '',
   title: '',
@@ -16,13 +21,11 @@ const ReviewForm = () => {
   const [previewImgs, setRreviewImgs] = useState([]);
   const [activedBtn, setActivedBtn] = useState('');
   const { title, contents } = review;
-  const fileRef = useRef();
-  const radioRef = useRef();
-
-  const handleImageUpload = (e) => {
+  const onImageUpload = (e) => {
     const files = e.target.files;
     if (files.length > 5) {
-      fileRef.current.value = '';
+      console.log(e.target);
+      e.target.value = '';
       alert('최대 5장까지 선택해주세요 : )');
       return;
     }
@@ -35,6 +38,10 @@ const ReviewForm = () => {
       let reader = new FileReader();
       reader.onload = () => {
         fileURLs[i] = reader.result;
+        // setReview({
+        //   ...review,
+        //   previewImgs: [...fileURLs],
+        // });
         setRreviewImgs([...fileURLs]);
       };
       reader.readAsDataURL(file);
@@ -45,7 +52,7 @@ const ReviewForm = () => {
     setActivedBtn(nextActivedBtn);
   };
 
-  const handleReviewChange = (e) => {
+  const onReviewChange = (e) => {
     const { name, value } = e.target;
     setReview({
       ...review,
@@ -60,54 +67,18 @@ const ReviewForm = () => {
     setRreviewImgs([]);
     setActivedBtn('');
     setReview(initReview);
-    fileRef.current.value = '';
-    radioRef.current.checked = false;
-    console.log(radioRef.current.checked);
   };
 
   return (
     <Wrap bg={BgImg}>
       <h2>피드 쓰기</h2>
       <form>
-        <Categoty>
-          <label>카테고리</label>
-          <input type="radio" name="category" value="stroe" onChange={handleReviewChange} ref={radioRef} /> 가게
-          <input type="radio" name="category" value="product" onChange={handleReviewChange} ref={radioRef} /> 제품
-        </Categoty>
-        <UploadImg>
-          <label>사진 선택 {previewImgs.length}/5</label>
-          <input type="file" accept="image/*" multiple onChange={handleImageUpload} ref={fileRef} />
-          <ul className="previewImgs">
-            <li className="imgContainer"></li>
-            <li className="imgContainer"></li>
-            <li className="imgContainer"></li>
-            <li className="imgContainer"></li>
-            <li className="imgContainer"></li>
-            {previewImgs?.map((img, index) => (
-              <li className="prevImg" key={img}>
-                <img src={img} alt="" index={index} />
-              </li>
-            ))}
-          </ul>
-        </UploadImg>
-        <VegLevel activedBtn={activedBtn} onReviewChange={handleReviewChange} onVegLevelClick={onVegLevelClick} />
-        <Title>
-          <label>제목</label>
-          <input name="title" value={title} placeholder="제목을 입력하세요." onChange={handleReviewChange}></input>
-        </Title>
-        <Contents>
-          <label>내용</label>
-          <textarea
-            name="contents"
-            value={contents}
-            placeholder="내용을 입력하세요."
-            onChange={handleReviewChange}></textarea>
-        </Contents>
-        <SubmitBtn>
-          <button type="submit" className="submitBtn" onClick={handleSubmit}>
-            피드 쓰기
-          </button>
-        </SubmitBtn>
+        <Category review={review} onReviewChange={onReviewChange} />
+        <UploadImg previewImgs={previewImgs} onImageUpload={onImageUpload} />
+        <VegLevel activedBtn={activedBtn} onReviewChange={onReviewChange} onVegLevelClick={onVegLevelClick} />
+        <Input label="제목" name="title" value={title} placeholder="제목을 입력하세요." onChange={onReviewChange} />
+        <Textarea name="contents" value={contents} placeholder="내용을 입력하세요." onChange={onReviewChange} />
+        <SubmitBtn value="피드 쓰기" onClick={handleSubmit} />
       </form>
     </Wrap>
   );
@@ -158,80 +129,5 @@ const Wrap = styled.section`
       background-image: url(${(props) => props.bg});
       background-size: cover;
     }
-  }
-`;
-
-const Categoty = styled.div`
-  input + input {
-    margin-left: 1rem;
-  }
-`;
-const UploadImg = styled.div`
-  .previewImgs {
-    position: relative;
-    width: 100%;
-    height: 150px;
-    margin-top: 1rem;
-    overflow: hidden;
-
-    input {
-      font-size: 0px;
-    }
-
-    .imgContainer {
-      border: 1px solid #999;
-    }
-    li,
-    .prevImg img {
-      float: left;
-      width: 150px;
-      height: 150px;
-      border-radius: 4px;
-    }
-    li + li {
-      margin-left: 0.5rem;
-    }
-    .prevImg img {
-      position: absolute;
-      top: 0;
-      left: ${(props) => props.index}px;
-      margin-left: -7px;
-    }
-  }
-`;
-const Title = styled.div`
-  input {
-    width: 100%;
-    padding: 10px;
-    border-radius: 4px;
-    border: 1px solid #999;
-  }
-`;
-const Contents = styled.div`
-  textarea {
-    width: 100%;
-    padding: 10px;
-    min-height: 500px;
-    border-radius: 4px;
-    border: 1px solid #999;
-  }
-`;
-const SubmitBtn = styled.div`
-  width: 33.3333%;
-  margin: 0 auto;
-  margin-top: 2rem;
-
-  border-radius: 4px;
-  border: 1px solid #999;
-  background-color: green;
-  text-align: center;
-
-  button {
-    width: 100%;
-    height: 100%;
-    padding: 0.65rem 0.75rem;
-    color: #fff;
-    -webkit-box-shadow: 0 3px 5px rgba(57, 63, 72, 0.25);
-    box-shadow: 0 3px 5px rgba(57, 63, 72, 0.25);
   }
 `;
